+++ services/streaming/app/repositories/tables/FilmTable.scala
@@ -11,7 +11,7 @@ class FilmTable(tag: Tag) extends Table[Film](tag, "films") {
   def id: Rep[Int]                   = column[Int]("id", O.PrimaryKey, O.AutoInc)
   def name: Rep[String]              = column[String]("name")
   def description: Rep[String]       = column[String]("description")
-  def duration: Rep[Int]             = column[Int]("duration")
+  def duration: Rep[Long]            = column[Long]("duration")
   def uploadDate: Rep[LocalDateTime] = column[LocalDateTime]("uploadDate")
   def views: Rep[Long]               = column[Long]("views")
   def available: Rep[Boolean]        = column[Boolean]("available")
